+++ deps/SConscript
@@ -5,8 +5,8 @@ libs_config = {
         "include": [
             'musl/src/internal',
             'musl/include',
-            'musl/arch/x86_64',
-            'musl/arch/x86_64/bits',
+            'musl/arch/'+ALT_ARCH,
+            'musl/arch/'+ALT_ARCH+'/bits',
             'libcxx/include',
             'eastl/include',
             'concurrentqueue',
@@ -392,30 +392,30 @@ libs_config = {
             'v8/src/wasm/wasm-opcodes.cc',
             'v8/src/wasm/wasm-result.cc',
             'v8/src/zone.cc',
-            'v8/src/crankshaft/x64/lithium-codegen-x64.cc',
-            'v8/src/crankshaft/x64/lithium-gap-resolver-x64.cc',
-            'v8/src/crankshaft/x64/lithium-x64.cc',
-            'v8/src/x64/assembler-x64.cc',
-            'v8/src/x64/builtins-x64.cc',
-            'v8/src/x64/code-stubs-x64.cc',
-            'v8/src/x64/codegen-x64.cc',
-            'v8/src/x64/cpu-x64.cc',
-            'v8/src/x64/deoptimizer-x64.cc',
-            'v8/src/x64/disasm-x64.cc',
-            'v8/src/x64/frames-x64.cc',
-            'v8/src/x64/interface-descriptors-x64.cc',
-            'v8/src/x64/macro-assembler-x64.cc',
-            'v8/src/debug/x64/debug-x64.cc',
-            'v8/src/full-codegen/x64/full-codegen-x64.cc',
-            'v8/src/ic/x64/access-compiler-x64.cc',
-            'v8/src/ic/x64/handler-compiler-x64.cc',
-            'v8/src/ic/x64/ic-x64.cc',
-            'v8/src/ic/x64/ic-compiler-x64.cc',
-            'v8/src/ic/x64/stub-cache-x64.cc',
-            'v8/src/regexp/x64/regexp-macro-assembler-x64.cc',
-            'v8/src/compiler/x64/code-generator-x64.cc',
-            'v8/src/compiler/x64/instruction-scheduler-x64.cc',
-            'v8/src/compiler/x64/instruction-selector-x64.cc',
+            'v8/src/crankshaft/'+ARCH+'/lithium-codegen-'+ARCH+'.cc',
+            'v8/src/crankshaft/'+ARCH+'/lithium-gap-resolver-'+ARCH+'.cc',
+            'v8/src/crankshaft/'+ARCH+'/lithium-'+ARCH+'.cc',
+            'v8/src/'+ARCH+'/assembler-'+ARCH+'.cc',
+            'v8/src/'+ARCH+'/builtins-'+ARCH+'.cc',
+            'v8/src/'+ARCH+'/code-stubs-'+ARCH+'.cc',
+            'v8/src/'+ARCH+'/codegen-'+ARCH+'.cc',
+            'v8/src/'+ARCH+'/cpu-'+ARCH+'.cc',
+            'v8/src/'+ARCH+'/deoptimizer-'+ARCH+'.cc',
+            'v8/src/'+ARCH+'/disasm-'+ARCH+'.cc',
+            'v8/src/'+ARCH+'/frames-'+ARCH+'.cc',
+            'v8/src/'+ARCH+'/interface-descriptors-'+ARCH+'.cc',
+            'v8/src/'+ARCH+'/macro-assembler-'+ARCH+'.cc',
+            'v8/src/debug/'+ARCH+'/debug-'+ARCH+'.cc',
+            'v8/src/full-codegen/'+ARCH+'/full-codegen-'+ARCH+'.cc',
+            'v8/src/ic/'+ARCH+'/access-compiler-'+ARCH+'.cc',
+            'v8/src/ic/'+ARCH+'/handler-compiler-'+ARCH+'.cc',
+            'v8/src/ic/'+ARCH+'/ic-'+ARCH+'.cc',
+            'v8/src/ic/'+ARCH+'/ic-compiler-'+ARCH+'.cc',
+            'v8/src/ic/'+ARCH+'/stub-cache-'+ARCH+'.cc',
+            'v8/src/regexp/'+ARCH+'/regexp-macro-assembler-'+ARCH+'.cc',
+            'v8/src/compiler/'+ARCH+'/code-generator-'+ARCH+'.cc',
+            'v8/src/compiler/'+ARCH+'/instruction-scheduler-'+ARCH+'.cc',
+            'v8/src/compiler/'+ARCH+'/instruction-selector-'+ARCH+'.cc',
 
             'v8/src/base/accounting-allocator.cc',
             'v8/src/base/atomicops_internals_x86_gcc.cc',
@@ -449,8 +449,8 @@ libs_config = {
         "include": [
             'musl/src/internal',
             'musl/include',
-            'musl/arch/x86_64',
-            'musl/arch/x86_64/bits',
+            'musl/arch/'+ALT_ARCH,
+            'musl/arch/'+ALT_ARCH+'/bits',
             'libcxx/include',
             '../src',
         ],
@@ -474,8 +474,8 @@ libs_config = {
         "include": [
             'musl/src/internal',
             'musl/include',
-            'musl/arch/x86_64',
-            'musl/arch/x86_64/bits',
+            'musl/arch/'+ALT_ARCH,
+            'musl/arch/'+ALT_ARCH+'/bits',
             'acpica/source/include',
         ],
         "source": [
@@ -1061,8 +1061,8 @@ libs_config = {
         "include": [
             'musl/src/internal',
             'musl/include',
-            'musl/arch/x86_64',
-            'musl/arch/x86_64/bits',
+            'musl/arch/'+ALT_ARCH,
+            'musl/arch/'+ALT_ARCH+'/bits',
         ],
         "source": [
             'libcxxrt/src/typeinfo.cc',
@@ -1074,8 +1074,8 @@ libs_config = {
         "include": [
             'musl/src/internal',
             'musl/include',
-            'musl/arch/x86_64',
-            'musl/arch/x86_64/bits',
+            'musl/arch/'+ALT_ARCH,
+            'musl/arch/'+ALT_ARCH+'/bits',
         ],
         "source": [
             'printf/printf.cc',
@@ -1085,8 +1085,8 @@ libs_config = {
         "include": [
             'musl/src/internal',
             'musl/include',
-            'musl/arch/x86_64',
-            'musl/arch/x86_64/bits',
+            'musl/arch/'+ALT_ARCH,
+            'musl/arch/'+ALT_ARCH+'/bits',
         ],
         "source": [
             'miniz/tinfl.c',
@@ -1096,8 +1096,8 @@ libs_config = {
         "include": [
             'musl/src/internal',
             'musl/include',
-            'musl/arch/x86_64',
-            'musl/arch/x86_64/bits',
+            'musl/arch/'+ALT_ARCH,
+            'musl/arch/'+ALT_ARCH+'/bits',
             'libsodium/src/libsodium/include/sodium',
         ],
         "source": [
@@ -1245,8 +1245,8 @@ libs_config = {
         "include": [
             'musl/src/internal',
             'musl/include',
-            'musl/arch/x86_64',
-            'musl/arch/x86_64/bits',
+            'musl/arch/'+ALT_ARCH,
+            'musl/arch/'+ALT_ARCH+'/bits',
             'libcxx/include',
             'json11'
         ],
