+++ src/kernel/arch/x86/startup_conf.inc
@@ -27,13 +27,13 @@ MULTIBOOT_LOAD_DATACODE_LEN = 0
 MULTIBOOT_LOAD_DATACODEBSS_LEN = 0
 
 ; SYSTEM VARS
-SYSTEM_IDT_TABLE_ADDR_64 = 0x0000000000000000
-SYSTEM_IDT_TABLE_ADDR_64_32len = 0x00000000 ; same but 32 bit
+SYSTEM_IDT_TABLE_ADDR_32 = 0x0000000000000000
+SYSTEM_IDT_TABLE_ADDR_32_32len = 0x00000000 ; same but 32 bit
 SYSTEM_IDT_COUNT = 256
 SYSTEM_IDT_LEN_BYTES = 16
 
-SYSTEM_GDT_TABLE_ADDR_64 = 0x0000000000001000
-SYSTEM_GDT_TABLE_ADDR_64_32len = 0x00001000	; same but 32 bit
+SYSTEM_GDT_TABLE_ADDR_32 = 0x0000000000001000
+SYSTEM_GDT_TABLE_ADDR_32_32len = 0x00001000	; same but 32 bit
 
 ; STACK AT 22 MiB
 SYSTEM_STACK_32BIT = 0x1600000
